Changed nelements to int64_t. Made the loop variables same across all three correlation functions. Added some colored output to examples code. Python bindings is still not functional

+++ xi_theory/wp/countpairs_wp.h
@@ -15,13 +15,13 @@ typedef struct{
 	int nbin;
 } results_countpairs_wp;
 
-results_countpairs_wp *countpairs_wp(const int ND1, const DOUBLE * restrict X1, const DOUBLE * restrict Y1, const DOUBLE * restrict Z1,
-																		 const double boxsize, 
+results_countpairs_wp *countpairs_wp(const int64_t ND1, const DOUBLE * restrict X1, const DOUBLE * restrict Y1, const DOUBLE * restrict Z1,
+									 const double boxsize, 
 #ifdef USE_OMP
-																		 const int numthreads,
+									 const int numthreads,
 #endif
-																		 const char *binfile,
-																		 const double pimax) __attribute__((warn_unused_result));
+									 const char *binfile,
+									 const double pimax) __attribute__((warn_unused_result));
 
 
 void free_results_wp(results_countpairs_wp **results);
