+++ xi_theory/wp/wp.c
@@ -15,9 +15,10 @@
 */
 
 #include <stdio.h>
-#include <math.h>
 #include <stdlib.h>
+#include <math.h>
 #include <assert.h>
+#include <inttypes.h>
 
 #include "defs.h" //for ADD_DIFF_TIME
 #include "function_precision.h" //definition of DOUBLE
@@ -39,7 +40,7 @@ int main(int argc, char *argv[])
   DOUBLE pimax ;
 	
   /*---Data-variables--------------------*/
-  int ND1=0,ND2=0;
+  int64_t ND1=0,ND2=0;
 
   DOUBLE *x1=NULL,*y1=NULL,*z1=NULL;
 
@@ -158,7 +159,7 @@ int main(int argc, char *argv[])
 	free_results_wp(&results);
 	
   gettimeofday(&t_end,NULL);
-  fprintf(stderr,"wp> Done -  ND1=%d ND2=%d. Time taken = %6.2lf seconds. read-in time = %6.2lf seconds pair-counting time = %6.2lf sec\n",
+  fprintf(stderr,"wp> Done -  ND1=%12"PRId64" ND2=%12"PRId64". Time taken = %6.2lf seconds. read-in time = %6.2lf seconds pair-counting time = %6.2lf sec\n",
 	  ND1,ND2,ADD_DIFF_TIME(t_start,t_end),read_time,pair_time);
   return EXIT_SUCCESS;
 }
