+++ xi_theory/xi_of_r/countpairs.h
@@ -14,13 +14,13 @@ typedef struct{
 	int nbin;
 } results_countpairs;
 
-results_countpairs * countpairs(const int ND1, const DOUBLE * const X1, const DOUBLE * const Y1, const DOUBLE  * const Z1,
-																const int ND2, const DOUBLE * const X2, const DOUBLE * const Y2, const DOUBLE  * const Z2,
+results_countpairs * countpairs(const int64_t ND1, const DOUBLE * const X1, const DOUBLE * const Y1, const DOUBLE  * const Z1,
+								const int64_t ND2, const DOUBLE * const X2, const DOUBLE * const Y2, const DOUBLE  * const Z2,
 #ifdef USE_OMP
-																const int numthreads,
+								const int numthreads,
 #endif
-																const int autocorr,
-																const char *binfile) __attribute__((warn_unused_result));
+								const int autocorr,
+								const char *binfile) __attribute__((warn_unused_result));
 
 
 void free_results(results_countpairs **results);
