+++ xi_theory/xi_rp_pi/DDrppi.c
@@ -15,9 +15,10 @@
 */
 
 #include <stdio.h>
-#include <math.h>
 #include <stdlib.h>
+#include <math.h>
 #include <assert.h>
+#include <inttypes.h>
 
 #include "defs.h" //for ADD_DIFF_TIME
 #include "function_precision.h" //definition of DOUBLE
@@ -39,7 +40,7 @@ int main(int argc, char *argv[])
 	
 	
   /*---Data-variables--------------------*/
-  int ND1=0,ND2=0;
+  int64_t ND1=0,ND2=0;
 
   DOUBLE *x1=NULL,*y1=NULL,*z1=NULL;
   DOUBLE *x2=NULL,*y2=NULL,*z2=NULL;//will point to x1/y1/z1 in case of auto-corr
@@ -160,13 +161,13 @@ int main(int argc, char *argv[])
 
 	const DOUBLE dpi = pimax/(DOUBLE)results->npibin ;
 	for(int i=1;i<results->nbin;i++) {
-		const double logrp = LOG10(results->rupp[i]);
-    for(int j=0;j<npibin;j++) {
+	  const double logrp = LOG10(results->rupp[i]);
+	  for(int j=0;j<npibin;j++) {
       int index = i*(npibin+1) + j;
 #ifdef OUTPUT_RPAVG			
       fprintf(stdout,"%10"PRIu64" %20.8lf %20.8lf  %20.8lf \n",results->npairs[index],results->rpavg[index],logrp,(j+1)*dpi);
 #else			
-			fprintf(stdout,"%10"PRIu64" %20.8lf %20.8lf  %20.8lf \n",results->npairs[index],0.0,logrp,(j+1)*dpi);
+	  fprintf(stdout,"%10"PRIu64" %20.8lf %20.8lf  %20.8lf \n",results->npairs[index],0.0,logrp,(j+1)*dpi);
 #endif			
     }
   }
@@ -175,7 +176,7 @@ int main(int argc, char *argv[])
 	free_results_rp_pi(&results);
 	
   gettimeofday(&t_end,NULL);
-  fprintf(stderr,"xi_rp_pi> Done -  ND1=%d ND2=%d. Time taken = %6.2lf seconds. read-in time = %6.2lf seconds pair-counting time = %6.2lf sec\n",
+  fprintf(stderr,"xi_rp_pi> Done -  ND1=%12"PRId64" ND2=%12"PRId64". Time taken = %6.2lf seconds. read-in time = %6.2lf seconds pair-counting time = %6.2lf sec\n",
 	  ND1,ND2,ADD_DIFF_TIME(t_start,t_end),read_time,pair_time);
   return EXIT_SUCCESS;
 }
